Fixes an inconsistency in cmake library consumption
When using include(<spzdir>) headers are expected to
be included as <spz-header.h>

When using find_package(spz) on the exported cmake-configuration
headers are expected to be included as <spz/spz-header.h>

This change makes all headers include consistently
in the second way by using target_include_directory
instead of relying on relative locations of headers
