+++ extensions/python/splat-extensions.cc
@@ -5,11 +5,11 @@
 #include <nanobind/stl/vector.h>
 #include <nanobind/stl/shared_ptr.h>
 
-#include "src/cc/load-spz.h"
-#include "src/cc/splat-types.h"
-#include "extensions/cc/splat-extensions.h"
-#include "extensions/cc/safe-orbit-camera-adobe.h"
-#include "extensions/cc/coordinate-system-adobe.h"
+#include <spz/load-spz.h>
+#include <spz/splat-types.h>
+#include <spz/splat-extensions.h>
+#include <spz/safe-orbit-camera-adobe.h>
+#include <spz/coordinate-system-adobe.h>
 #include "extensions/python/splat-extensions.h"
 
 namespace nb = nanobind;
