+++ include/spz/safe-orbit-camera-adobe.h
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2026 Adobe Inc.
+Copyright (c) 2025 Adobe Inc.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -22,32 +22,21 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#ifndef SPZ_COORDINATE_SYSTEM_ADOBE_H_
-#define SPZ_COORDINATE_SYSTEM_ADOBE_H_
+#ifndef SPZ_SAFE_ORBIT_CAMERA_ADOBE_H_
+#define SPZ_SAFE_ORBIT_CAMERA_ADOBE_H_
 
-#include "splat-extensions.h"
-#include "splat-types.h"
+#include <spz/splat-extensions.h>
 
 namespace spz {
 
-// Records the coordinate system in which the Gaussian data is physically stored in the file.
-//
-// This is a descriptor, not an instruction. It labels what coordinate system the packed data is
-// already in — it does not tell the reader to apply any conversion. packGaussians and
-// unpackGaussians consume this extension internally and handle all conversions automatically via
-// PackOptions::from and UnpackOptions::to. Consumers using those functions must not apply any
-// additional conversion based on this extension; doing so will double-transform the data.
-//
-// Mechanics: when present with a non-UNSPECIFIED value, packGaussians writes data in
-// coordinateSystem (instead of the default RUB), and unpackGaussians reads from coordinateSystem
-// (instead of RUB) before converting to UnpackOptions::to.
-//
-// UNSPECIFIED is equivalent to the extension being absent: both fall back to RUB, but a present
-// UNSPECIFIED value is likely a writer bug and will produce a warning.
-struct SpzExtensionCoordinateSystemAdobe : public SpzExtensionBase {
-  CoordinateSystem coordinateSystem = CoordinateSystem::UNSPECIFIED;
+struct SpzExtensionSafeOrbitCameraAdobe : public SpzExtensionBase {
+  float safeOrbitElevationMin = 0.0f;  // Minimum elevation for safe orbit (radians)
+  float safeOrbitElevationMax = 0.0f;  // Maximum elevation for safe orbit (radians)
+  float safeOrbitRadiusMin = 0.0f;     // Minimum radius for safe orbit
 
-  SpzExtensionCoordinateSystemAdobe();
+  static const std::unordered_set<std::string> kRequiredPlyElementNames;
+
+  SpzExtensionSafeOrbitCameraAdobe();
   uint32_t payloadBytes() const override;
   void write(std::ostream& os) const override;
   SpzExtensionBase* copyAsRawData() const override;
@@ -57,12 +46,8 @@ struct SpzExtensionCoordinateSystemAdobe : public SpzExtensionBase {
   void writePlyData(std::ostream& out) const override;
   static std::optional<SpzExtensionBasePtr> read(std::istream& is);
   static SpzExtensionType type();
-
-  // Returns the storage coordinate system recorded by this extension.
-  // Returns RUB and logs a warning if coordinateSystem is UNSPECIFIED (likely a writer bug).
-  CoordinateSystem resolve() const;
 };
 
 }  // namespace spz
 
-#endif  // SPZ_COORDINATE_SYSTEM_ADOBE_H_
+#endif  // SPZ_SAFE_ORBIT_CAMERA_ADOBE_H_
