+++ include/spz/splat-extensions.h
