+++ src/cc/splat-c-types.cc
@@ -22,6 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include "splat-c-types.h"
+#include <spz/splat-c-types.h>
 
 // Empty.
