+++ src/cc/splat-types.cc
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include "splat-types.h"
+#include <spz/splat-types.h>
 
 #include <cmath>
 #include <limits>
