+++ src/cc/splat-utils.h
@@ -4,8 +4,8 @@
 #include <cmath>
 #include <cstdint>
 
-#include "load-spz.h"
-#include "splat-types.h"
+#include <spz/load-spz.h>
+#include <spz/splat-types.h>
 
 namespace spz {
